+++ src/main/java/gtclassic/api/helpers/GTUtility.java
@@ -2,7 +2,10 @@ package gtclassic.api.helpers;
 
 import java.util.List;
 
+import gtclassic.common.tile.GTTileMagicEnergyAbsorber;
+import gtclassic.common.util.GTIBlockFilters;
 import gtclassic.common.util.GTIFilters;
+import ic2.core.RotationList;
 import ic2.core.block.base.tile.TileEntityMachine;
 import ic2.core.fluid.IC2Tank;
 import ic2.core.inventory.filters.CommonFilters;
@@ -10,7 +13,10 @@ import ic2.core.inventory.filters.IFilter;
 import ic2.core.inventory.transport.IItemTransporter;
 import ic2.core.inventory.transport.TransporterManager;
 import ic2.core.item.armor.electric.ItemArmorQuantumSuit;
+import ic2.core.util.helpers.AabbUtil;
 import ic2.core.util.math.MathUtil;
+import net.minecraft.block.BlockEndPortalFrame;
+import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLiving;
 import net.minecraft.entity.EntityLivingBase;
@@ -18,8 +24,12 @@ import net.minecraft.entity.IProjectile;
 import net.minecraft.entity.item.EntityArmorStand;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.projectile.EntityArrow;
+import net.minecraft.init.Blocks;
+import net.minecraft.init.SoundEvents;
 import net.minecraft.item.ItemStack;
+import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
+import net.minecraft.util.SoundCategory;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Vec3d;
@@ -165,5 +175,57 @@ public class GTUtility {
 		if (canExport) {
 			FluidUtil.tryFluidTransfer(fluidTile, tank, amount, true);
 		}
+	}
+
+	public static boolean tryResetStrongholdPortal(World world, BlockPos pos) {
+		List<BlockPos> portalBlockPos = AabbUtil.getTargets(world, pos, 5, new GTIBlockFilters.EndPortalFilter(), false, false, RotationList.ALL);
+		if (portalBlockPos.isEmpty()) {
+			return false;
+		}
+		BlockPos selectedPos = portalBlockPos.get(world.rand.nextInt(portalBlockPos.size() - 1));
+		IBlockState nearbyState = world.getBlockState(selectedPos);
+		if (resetEndPortalFrame(world, selectedPos, nearbyState)) {
+			boolean found = false;
+			for (BlockPos portalPos : portalBlockPos) {
+				if (world.getBlockState(portalPos).getBlock() == Blocks.END_PORTAL) {
+					world.setBlockToAir(portalPos);
+					world.removeTileEntity(portalPos);
+					found = true;
+				}
+			}
+			return found;
+		}
+		return false;
+	}
+
+	public static boolean resetEndPortalFrame(World world, BlockPos pos, IBlockState portalFrameState) {
+		if (portalFrameState.getBlock() == Blocks.END_PORTAL_FRAME
+				&& portalFrameState.getValue(BlockEndPortalFrame.EYE).booleanValue()) {
+			world.setBlockState(pos, portalFrameState.withProperty(BlockEndPortalFrame.EYE, false));
+			world.playSound((EntityPlayer) null, pos, SoundEvents.BLOCK_END_PORTAL_FRAME_FILL, SoundCategory.BLOCKS, 0.5F, 0.5F
+					+ world.rand.nextFloat());
+			return true;
+		}
+		return false;
+	}
+
+	public static boolean tryExplodeOtherAbsorbers(World world, BlockPos checkPos) {
+		Iterable<BlockPos> surroundingPos = BlockPos.getAllInBox(checkPos.offset(EnumFacing.SOUTH, 4).offset(EnumFacing.WEST, 4), checkPos.offset(EnumFacing.NORTH, 4).offset(EnumFacing.EAST, 4));
+		for (BlockPos absorberPos : surroundingPos) {
+			if (absorberPos.equals(checkPos)) {
+				continue;
+			}
+			TileEntity tile = world.getTileEntity(absorberPos);
+			if (tile instanceof GTTileMagicEnergyAbsorber) {
+				GTTileMagicEnergyAbsorber absorber = (GTTileMagicEnergyAbsorber) tile;
+				if (absorber.portalMode && absorber.isAbovePortal) {
+					world.setBlockToAir(absorberPos);
+					world.removeTileEntity(absorberPos);
+					world.createExplosion(null, absorberPos.getX(), absorberPos.getY(), absorberPos.getZ(), 8.0F, true);
+					return true;
+				}
+			}
+		}
+		return false;
 	}
 }