+++ src/main/java/gtclassic/common/util/GTIBlockFilters.java
@@ -1,12 +1,12 @@
 package gtclassic.common.util;
 
-import gtclassic.api.interfaces.IGTDataNetObject;
 import gtclassic.common.GTBlocks;
 import gtclassic.common.tile.multi.GTTileMultiLESU;
 import ic2.core.util.helpers.AabbUtil;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.init.Blocks;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
@@ -92,26 +92,22 @@ public class GTIBlockFilters {
 		}
 	}
 
-	public static class DataNetFilter implements AabbUtil.IBlockFilter {
-
-		/** This filter will loop over ANY valid data net block **/
-		public DataNetFilter() {
-		}
-
-		@Override
-		public boolean isValidBlock(IBlockState state) {
-			return state.getBlock() instanceof IGTDataNetObject;
-		}
-
-		@Override
-		public boolean isValidBlock(World world, BlockPos blockPos) {
-			Block worldBlock = world.getBlockState(blockPos).getBlock();
-			return worldBlock instanceof IGTDataNetObject || world.getTileEntity(blockPos) instanceof IGTDataNetObject;
-		}
+	public static class EndPortalFilter implements AabbUtil.IBlockFilter {
 
 		@Override
 		public boolean forceChunkLoad() {
 			return false;
+		}
+
+		@Override
+		public boolean isValidBlock(IBlockState state) {
+			return state.getBlock() == Blocks.END_PORTAL || state.getBlock() == Blocks.END_PORTAL_FRAME;
+		}
+
+		@Override
+		public boolean isValidBlock(World world, BlockPos pos) {
+			return world.getBlockState(pos).getBlock() == Blocks.END_PORTAL
+					|| world.getBlockState(pos).getBlock() == Blocks.END_PORTAL_FRAME;
 		}
 	}
 }